duble to Bigdecimal2

+++ src/main/java/sk/zpn/zaklad/model/util/TestDataFactory.java
@@ -4,6 +4,7 @@ import com.vaadin.server.VaadinSession;
 import sk.zpn.domena.Produkt;
 
 import javax.persistence.EntityManager;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,30 +19,30 @@ public class TestDataFactory {
     public void createProdukty() {
         List<Produkt> products = new ArrayList<>();
         products.add((Produkt) new Produkt()
-            .setBody(1D)
+            .setBody(BigDecimal.valueOf(1))
             .setKat("01kJG")
-            .setKusy(1D)
+            .setKusy(BigDecimal.valueOf(1))
             .setNazov("baban")
             .setRok("2019")
             .setId(0L));
         products.add((Produkt) new Produkt()
-            .setBody(1.3D)
+            .setBody(BigDecimal.valueOf(1.3))
             .setKat("02YSH")
-            .setKusy(5D)
+            .setKusy(BigDecimal.valueOf(5))
             .setNazov("zemiak")
             .setRok("2019")
             .setId(0L));
         products.add((Produkt) new Produkt()
-            .setBody(4D)
+            .setBody(BigDecimal.valueOf(4))
             .setKat("03UFE")
-            .setKusy(1D)
+            .setKusy(BigDecimal.valueOf(1))
             .setNazov("ananas")
             .setRok("2019")
             .setId(0L));
         products.add((Produkt) new Produkt()
-            .setBody(2D)
+            .setBody(BigDecimal.valueOf(2))
             .setKat("04UHF")
-            .setKusy(3D)
+            .setKusy(BigDecimal.valueOf(3))
             .setNazov("pomelo")
             .setRok("2019")
             .setId(0L));
