+++ src/main/java/sk/zpn/zaklad/view/produkty/BrowsPanel.java
@@ -8,6 +8,7 @@ import sk.zpn.domena.Produkt;
 import sk.zpn.domena.Uzivatel;
 import sk.zpn.zaklad.view.VitajteView;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.function.Consumer;
 
@@ -36,8 +37,8 @@ public class BrowsPanel extends VerticalLayout {
             FilterGrid.Column<Produkt, String > colRok = grid.addColumn(Produkt::getRok).setCaption("Rok").setId("rok");
             FilterGrid.Column<Produkt, String> colKat = grid.addColumn(Produkt::getKat).setCaption("KAT").setId("kat");
             FilterGrid.Column<Produkt, String> colNazov = grid.addColumn(Produkt::getNazov).setCaption("Názov").setId("nazov");
-            FilterGrid.Column<Produkt, Double> colKusy = grid.addColumn(Produkt::getKusy).setCaption("kusy").setId("kusy");
-            FilterGrid.Column<Produkt, Double> colBody = grid.addColumn(Produkt::getBody).setCaption("Body").setId("body");
+            FilterGrid.Column<Produkt, BigDecimal> colKusy = grid.addColumn(Produkt::getKusy).setCaption("kusy").setId("kusy");
+            FilterGrid.Column<Produkt, BigDecimal> colBody = grid.addColumn(Produkt::getBody).setCaption("Body").setId("body");
 //            TODO pridat cez query na vazobnu entitu
 //            FilterGrid.Column<Produkt, String> colFirmaNazov = grid.addColumn(Produkt::getFirmaNazov).setCaption("Firma").setId("nazovFirmy");
